+++ Tests/TronTestCase.swift
@@ -9,16 +9,57 @@
 import XCTest
 import TRON
 import Nimble
+import Alamofire
 
 class TronTestCase: XCTestCase {
     
+    var tron: TRON!
+    
+    override func setUp() {
+        super.setUp()
+        tron = TRON(baseURL: "https://github.com")
+    }
+    
     func testTronRequestBuildables() {
-        let tron = TRON(baseURL: "https://github.com")
         let request: APIRequest<Int,TronError> = tron.request(path: "/foo")
         
         let tronBuilder = tron.urlBuilder as? URLBuilder
         let requestBuilder = request.urlBuilder as? URLBuilder
         expect(requestBuilder === tronBuilder).to(beTruthy())
+    }
+    
+    func testURLEncodingStrategySetsURLEncoding() {
+        tron.encodingStrategy = TRON.URLEncodingStrategy()
+        
+        let request : APIRequest<Int,TronError> = tron.request(path: "foo")
+        request.method = .POST
+        
+        if case ParameterEncoding.URL = request.encoding {
+            
+        } else {
+            XCTFail()
+        }
+    }
+    
+    func testRESTEncodingStrategySetsProperEncoding() {
+        tron.encodingStrategy = TRON.RESTEncodingStrategy()
+        
+        let request : APIRequest<Int,TronError> = tron.request(path: "foo")
+        request.method = .POST
+        
+        if case ParameterEncoding.JSON = request.encoding {
+        
+        } else {
+            XCTFail()
+        }
+        
+        request.method = .GET
+        
+        if case ParameterEncoding.URL = request.encoding {
+            
+        } else {
+            XCTFail()
+        }
     }
     
 }